+++ backend/src/main/java/com/neoadventura/dtos/CreateUsuarioDto.java
@@ -2,6 +2,7 @@ package com.neoadventura.dtos;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Getter
@@ -12,8 +13,8 @@ public class CreateUsuarioDto {
     private Date birth_day;
     private Date registered;
     private Boolean suscribed;
-    private Double monedero_virtual;
-    private Double monedero_oferta;
+    private BigDecimal monedero_virtual;
+    private BigDecimal monedero_oferta;
     private Boolean same_language;
     private Boolean banned;
     private Long rol_id;
