+++ backend/src/main/java/com/neoadventura/dtos/ServicioDto.java
@@ -3,6 +3,8 @@ package com.neoadventura.dtos;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+import java.util.Currency;
 import java.util.Date;
 
 @Getter
@@ -13,7 +15,7 @@ public class ServicioDto {
     private String description;
     private Date init_valid_date;
     private Date end_valid_date;
-    private Double price;
+    private BigDecimal price;
     private Long modalidad_id;
     private Long region_id;
     private Long plataforma_id;
