+++ backend/src/main/java/com/neoadventura/dtos/UsuarioDto.java
@@ -3,6 +3,7 @@ package com.neoadventura.dtos;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Getter
@@ -15,8 +16,8 @@ public class UsuarioDto {
     private Date birth_day;
     private Date registered;
     private Boolean suscribed;
-    private Double monedero_virtual;
-    private Double monedero_oferta;
+    private BigDecimal monedero_virtual;
+    private BigDecimal monedero_oferta;
     private Boolean same_language;
     private Boolean banned;
     private Long rol_id;
